fix: refactor playout-gateway types into shared-lib
