+++ packages/blueprints-integration/package.json
@@ -39,7 +39,6 @@
 	],
 	"dependencies": {
 		"@sofie-automation/shared-lib": "link:../shared-lib",
-		"timeline-state-resolver-types": "8.0.0-release48.0",
 		"tslib": "^2.4.0",
 		"type-fest": "^2.19.0"
 	},
