+++ packages/playout-gateway/src/tsrHandler.ts
@@ -4,9 +4,7 @@ import {
 	ConductorOptions,
 	TimelineTriggerTimeResult,
 	DeviceOptionsAny,
-	Mappings,
 	DeviceContainer,
-	Timeline as TimelineTypes,
 	TSRTimelineObj,
 	TSRTimeline,
 	TSRTimelineContent,
@@ -40,8 +38,13 @@ import {
 	PeripheralDeviceStatusObject,
 } from '@sofie-automation/shared-lib/dist/peripheralDevice/peripheralDeviceAPI'
 import { assertNever } from '@sofie-automation/shared-lib/dist/lib/lib'
-import { protectString } from '@sofie-automation/shared-lib/dist/lib/protectedString'
+import { protectString, unprotectObject } from '@sofie-automation/shared-lib/dist/lib/protectedString'
 import { StudioId, TimelineHash } from '@sofie-automation/shared-lib/dist/core/model/Ids'
+import {
+	RoutedMappings,
+	RoutedTimeline,
+	TimelineObjGeneric,
+} from '@sofie-automation/shared-lib/dist/core/model/Timeline'
 import {
 	PeripheralDevicePublic,
 	PlayoutDeviceSettings,
@@ -53,55 +56,6 @@ const debug = Debug('playout-gateway')
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface TSRConfig {}
 
-// ----------------------------------------------------------------------------
-// interface copied from Core lib/collections/Timeline.ts
-
-export type TimelineEnableExt = TimelineTypes.TimelineEnable & { setFromNow?: boolean }
-export interface TimelineObjGeneric extends TSRTimelineObj<any> {
-	/** Unique within a timeline (ie within a studio) */
-	id: string
-	/** Set when the id of the object is prefixed */
-	originalId?: string
-
-	objectType: TimelineObjType
-
-	enable: TimelineEnableExt | TimelineEnableExt[]
-
-	/** The id of the group object this object is in  */
-	inGroup?: string
-}
-export enum TimelineObjType {
-	/** Objects played in a rundown */
-	RUNDOWN = 'rundown',
-	/** Objects controlling recording */
-	RECORDING = 'record',
-	/** Objects controlling manual playback */
-	MANUAL = 'manual',
-	/** "Magic object", used to calculate a hash of the timeline */
-	STAT = 'stat',
-}
-
-/** This is the data-object published from Core */
-export interface RoutedTimeline {
-	_id: StudioId
-	/** Hash of the studio mappings */
-	mappingsHash: string
-
-	/** Hash of the Timeline */
-	timelineHash: string
-
-	/** serialized JSON Array containing all timeline-objects */
-	timelineBlob: string
-	generated: number
-
-	// this is the old way of storing the timeline, kept for backwards-compatibility
-	timeline?: TimelineObjGeneric[]
-}
-export interface RoutedMappings {
-	_id: StudioId
-	mappingsHash: string | undefined
-	mappings: Mappings
-}
 // ----------------------------------------------------------------------------
 
 export interface TimelineContentObjectTmp<TContent extends { deviceType: DeviceType }>
@@ -447,7 +401,7 @@ export class TSRHandler {
 			? this._transformTimeline(clone(timeline.timeline))
 			: []
 		this.tsr.timelineHash = timeline.timelineHash
-		this.tsr.setTimelineAndMappings(transformedTimeline, mappingsObject.mappings)
+		this.tsr.setTimelineAndMappings(transformedTimeline, unprotectObject(mappingsObject.mappings))
 	}
 	private _getPeripheralDevice(): PeripheralDevicePublic {
 		const peripheralDevices = this._coreHandler.core.getCollection<PeripheralDevicePublic>('peripheralDevices')
@@ -1057,44 +1011,33 @@ export class TSRHandler {
 	 * @param timeline
 	 */
 	private _transformTimeline(timeline: Array<TimelineObjGeneric>): TSRTimeline {
-		const transformObject = (obj: TimelineObjGeneric): TimelineContentObjectTmp<TSRTimelineContent> => {
-			if (!obj.content) obj.content = {}
-			if (obj.isGroup) {
-				if (!obj.content.objects) obj.content.objects = []
-			}
-
-			return obj
-		}
-
 		// First, transform and convert timeline to a key-value store, for fast referencing:
 		const objects: { [id: string]: TimelineContentObjectTmp<TSRTimelineContent> } = {}
-		_.each(timeline, (obj: TimelineObjGeneric) => {
-			const transformedObj = transformObject(obj)
-			objects[transformedObj.id] = transformedObj
-		})
+		for (const obj of timeline) {
+			objects[obj.id] = obj
+		}
 
 		// Go through all objects:
 		const transformedTimeline: Array<TSRTimelineObj<TSRTimelineContent>> = []
-		_.each(objects, (obj: TimelineContentObjectTmp<TSRTimelineContent>) => {
+		for (const obj of Object.values(objects)) {
 			if (!obj.inGroup) {
 				// Add object to timeline
 				delete obj.inGroup
 				transformedTimeline.push(obj)
-				return
+				continue
 			}
 			const groupObj = objects[obj.inGroup]
 			if (!groupObj) {
 				// referenced group not found
 				this.logger.error(`Referenced group "${obj.inGroup}" not found! Referenced by "${obj.id}"`)
-				return
+				continue
 			}
 			// Add object into group:
 			if (!groupObj.children) groupObj.children = []
-			if (groupObj.children) {
-				delete obj.inGroup
-				groupObj.children.push(obj)
-			}
-		})
+			groupObj.children.push(obj)
+			delete obj.inGroup
+		}
+
 		return transformedTimeline
 	}
