+++ packages/shared-lib/package.json
@@ -39,6 +39,7 @@
 		"/LICENSE"
 	],
 	"dependencies": {
+		"timeline-state-resolver-types": "8.0.0-release48.0",
 		"tslib": "^2.4.0",
 		"type-fest": "^2.19.0"
 	},
