+++ packages/shared-lib/src/tsr.ts
@@ -1,0 +1,5 @@
+import * as TSR from 'timeline-state-resolver-types'
+export { TSR }
+
+import * as tsrPkgInfo from 'timeline-state-resolver-types/package.json'
+export const TSR_VERSION: string = tsrPkgInfo.version
